requirements.txt
Flask
gunicorn
sentence-transformers
faiss-cpu
openai
python-dotenv